requirements.txt
gradio
speechrecognition
PyMuPDF
transformers
torch
faiss-cpu
numpy
gtts
pydub
groq
python-dotenv